Edit: Create separate folder for assets
