+++ src/pages/property/PropertyEditForm.js
@@ -13,7 +13,7 @@ import appStyles from "../../App.module.css";
 import btnStyles from "../../styles/Button.module.css";
 import styles from "../../styles/PropertyCreateEditForm.module.css";
 import Upload from "../../assets/upload.png";
-import Asset from "../../components/Asset";
+import Asset from "../../components/asset/Asset.js";
 
 import { useRedirect } from "../../hooks/useRedirect";
 import { axiosReq } from "../../api/axiosDefaults";
